place session_id in utils.py
